Update: Rename stream.js to readable.js to align with docs
